+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   - bump out `.string()` method deprecation removals to `0.0.61`
 - `ryo3-std`
   - `ry.FileType` struct now has inner faux type/real type
+- `ryo3-tokio`
+  - renamed:
+    - `DirEntryAsync` -> `AsyncDirEntry`
+    - `ReadDirAsync` -> `AsyncReadDir`
 
 ---